Add a Dockerfile to this repository
This commit adds a two-stage Docker build to this repository. The first
step uses an Alpine Linux based image to build the source of jo, and
then this stage is discarded (except for the compiled binary) and a new
Alpine based image is produced containing the jo binary.

The built image can then be executed as:

    [user@host] $ docker run --rm -ti jo a=b
    {"a": "b"}