Support reading nested data from pipes
slurp_file was adapted so that if it cannot determine the size of the
input file, the buffer it reads into is dynamically extended while
reading.

Input files may now be specified as "-" to read from jo's standard
input.

slurp_file's signature was changed so that the name of the file to
read is passed as argument, rather than a file pointer, as the file
pointer was never used after the file was slurped anyway.